📝 Docs : 알림 문서 작성 및 빌드
