+++ Backend/pom.xml
@@ -10,7 +10,7 @@
 	</parent>
 	<groupId>kimjang.toolkit</groupId>
 	<artifactId>solsol</artifactId>
-	<version>0.2.1</version>
+	<version>0.3.1</version>
 	<name>solsol</name>
 	<description>develop chatting server</description>
 	<properties>
