+++ Backend/src/main/java/kimjang/toolkit/solsol/room/service/ChatRoomStompService.java
@@ -25,6 +25,7 @@ public class ChatRoomStompService {
                             .roomName(withParticipationsName(createChatRoomDto, customerDto.getId()))
                             .firstChat(createChatRoomDto.getFirstChat())
                             .customer(customerDto)
+                            .maker(createChatRoomDto.getMaker())
                             .build();
                     messagingTemplate.convertAndSend("/notification/room/" + customerDto.getId(), // 각 고객에게 채팅방 생성을 알림
                             createRoomReqDto);
